tools/GRC_automation_scripts/compliance_ledger/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
PyYAML>=6.0
cryptography>=3.4.8